// web/src/api/generated/models/grading-result-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * NovaCloudEdu API
 * NovaCloudEdu Smart Education Platform Backend API
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { QuestionGradingItem } from './question-grading-item';

/**
 * 批改结果响应
 */
export interface GradingResultResponse {
    /**
     * 提交ID
     */
    'submissionId'?: string;
    /**
     * 总得分
     */
    'totalScore'?: number;
    /**
     * 满分
     */
    'maxScore'?: number;
    /**
     * 总评语
     */
    'overallComment'?: string;
    /**
     * 使用的AI模型
     */
    'modelId'?: string;
    /**
     * 批改完成时间
     */
    'gradingTime'?: string;
    /**
     * 逐题批改详情
     */
    'questions'?: Array<QuestionGradingItem>;
}