// web/src/api/generated/models/question-grading-item.ts
/* tslint:disable */
/* eslint-disable */
/**
 * NovaCloudEdu API
 * NovaCloudEdu Smart Education Platform Backend API
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 单题批改详情
 */
export interface QuestionGradingItem {
    /**
     * 题号
     */
    'questionIndex'?: number;
    /**
     * 题干
     */
    'questionContent'?: string;
    /**
     * 题型
     */
    'questionType'?: string;
    /**
     * 学生答案
     */
    'studentAnswer'?: string;
    /**
     * 标准答案
     */
    'standardAnswer'?: string;
    /**
     * 得分
     */
    'score'?: number;
    /**
     * 满分
     */
    'maxScore'?: number;
    /**
     * 错误分类
     */
    'errorCategories'?: Array<string>;
    /**
     * 错误详情
     */
    'errorDetail'?: string;
    /**
     * 关联知识点
     */
    'knowledgePoints'?: Array<string>;
    /**
     * 评语
     */
    'comment'?: string;
}